requirements.txt
Flask==2.0.2
python-dotenv==0.14.0
autopep8
pylint
Flask_SQLAlchemy==2.5.1
Flask-Migrate
psycopg2
Flask-Script==2.0.6
flask_marshmallow==0.14.0